+++ CMakeLists.txt
@@ -22,6 +22,7 @@ add_library(CubicScript STATIC
     "src/interpreter/interpreter.c"
     "src/interpreter/function_definition.c"
     "src/interpreter/operations.c"
+    "src/interpreter/stack.c"
     "src/sync/atomic.c"
     "src/sync/locks.c"
     "src/sync/sync_queue.c"
