+++ src/tests.zig
@@ -4,6 +4,7 @@ comptime {
 
     _ = @import("program/program.zig").Program;
     _ = @import("interpreter/interpreter.zig");
+    _ = @import("interpreter/stack.zig");
     _ = @import("interpreter/function_definition.zig");
     _ = @import("program/protected_arena.zig");
